Refactor to use the inputbar as a UIResponder inputAccessoryView.
This isn’t entirely future-proof for changes that might be made upstream in
SlackTextViewController later on. It probably requires some further internal
design changes to be clean enough to be merged by upstream.

This also breaks the pan to dismiss gesture. Not entirely important yet as it
doesn’t work on iOS 9 anyway.